Moved more tests
