+++ tests/unit/controllers/test_collection_instrument_controller.py
