+++ tests/unit/controllers/test_party_controller.py
